% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_document.R
\name{check_document}
\alias{check_document}
\title{Check current document}
\usage{
check_document(ask_modify = TRUE, language = "en-US")
}
\arguments{
\item{ask_modify}{logical}

\item{language}{en-US, FR, etc}
}
\description{
Check current source document.
}
