% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_selection.R
\name{check_selection}
\alias{check_selection}
\title{Check current selection}
\usage{
check_selection(ask_modify = TRUE, language = NULL)
}
\arguments{
\item{ask_modify}{logical. When \code{TRUE}, a prompt will be shown asking to alter the document with the correction or not.}

\item{language}{character. Code like en-US, fr-FR, etc}
}
\value{
No object is return from this function, as it expects to be run in an interactive session.
}
\description{
Check for grammar errors on the selected text and recommend corrections when possible.
}
