% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spell_check.R
\name{check_text}
\alias{check_text}
\title{Grammar check a text}
\usage{
check_text(text, language = "auto")
}
\arguments{
\item{text}{character. A character object with text to be check.}

\item{language}{character. Code like \code{en-US}, \code{fr-FR}, etc. Or \code{auto} (default) for automatic language detection.}
}
\value{
a tibble with grammar checks.
}
\description{
Grammar check a text
}
\examples{
check_text("O rato roeu a roupa do rei de roma.", language = "pt-BR")
check_text("Bernardo climb the stairs to the castle’s ramparts.", language = "auto")
}
