% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_languages.R
\name{get_languages}
\alias{get_languages}
\title{Get supported languages from LanguageTool API}
\usage{
get_languages()
}
\value{
a tibble.
}
\description{
Returns a tibble of supported languages from LanguageTools API, with language name, code and long code.
}
\examples{
\dontshow{if (curl::has_internet()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_languages()
\dontshow{\}) # examplesIf}
}
