% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_selection.R
\name{check_selection}
\alias{check_selection}
\title{Check current selection}
\usage{
check_selection(ask_modify = TRUE, language = "en-US")
}
\arguments{
\item{ask_modify}{logical}

\item{language}{en-US, FR, etc}
}
\value{
nothing.
}
\description{
Check current selection
}
