% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_selection.R
\name{check_selection}
\alias{check_selection}
\title{Check current selection}
\usage{
check_selection(ask_modify = TRUE, language = NULL)
}
\arguments{
\item{ask_modify}{logical.}

\item{language}{code like en-US, fr-FR, etc}
}
\value{
nothing.
}
\description{
Check for grammar errors at the current document text selection and provide alterations suggestions if available.
}
